+++ src/snyk/base/modules/baseSnykModule.ts
@@ -1,14 +1,16 @@
 import { AdvisorApiClient, IAdvisorApiClient } from '../../advisor/services/advisorApiClient';
 import AdvisorProvider from '../../advisor/services/advisorProvider';
 import { AdvisorService } from '../../advisor/services/advisorService';
-import { DownloadService } from '../../common/services/downloadService';
 import { IAnalytics } from '../../common/analytics/itly';
 import { ISnykApiClient, SnykApiClient } from '../../common/api/apiСlient';
 import { CommandController } from '../../common/commands/commandController';
 import { configuration } from '../../common/configuration/instance';
+import { IWorkspaceTrust, WorkspaceTrust } from '../../common/configuration/trustedFolders';
 import { ExperimentService } from '../../common/experiment/services/experimentService';
+import { ILanguageServer } from '../../common/languageServer/languageServer';
 import { Logger } from '../../common/logger/logger';
 import { ContextService, IContextService } from '../../common/services/contextService';
+import { DownloadService } from '../../common/services/downloadService';
 import { LearnService } from '../../common/services/learnService';
 import { INotificationService } from '../../common/services/notificationService';
 import { IOpenerService, OpenerService } from '../../common/services/openerService';
@@ -19,6 +21,7 @@ import { IMarkdownStringAdapter, MarkdownStringAdapter } from '../../common/vsco
 import { vsCodeWorkspace } from '../../common/vscode/workspace';
 import { IWatcher } from '../../common/watchers/interfaces';
 import { ISnykCodeService } from '../../snykCode/codeService';
+import { ISnykCodeServiceOld } from '../../snykCode/codeServiceOld';
 import { CodeSettings, ICodeSettings } from '../../snykCode/codeSettings';
 import { ISnykCodeErrorHandler, SnykCodeErrorHandler } from '../../snykCode/error/snykCodeErrorHandler';
 import { FalsePositiveApi, IFalsePositiveApi } from '../../snykCode/falsePositive/api/falsePositiveApi';
@@ -30,7 +33,6 @@ import { ScanModeService } from '../services/scanModeService';
 import SnykStatusBarItem, { IStatusBarItem } from '../statusBarItem/statusBarItem';
 import { ILoadingBadge, LoadingBadge } from '../views/loadingBadge';
 import { IBaseSnykModule } from './interfaces';
-import { ILanguageServer } from '../../common/languageServer/languageServer';
 
 export default abstract class BaseSnykModule implements IBaseSnykModule {
   context: ExtensionContext;
@@ -60,6 +62,7 @@ export default abstract class BaseSnykModule implements IBaseSnykModule {
   protected snykApiClient: ISnykApiClient;
   protected advisorApiClient: IAdvisorApiClient;
   protected falsePositiveApi: IFalsePositiveApi;
+  snykCodeOld: ISnykCodeServiceOld;
   snykCode: ISnykCodeService;
   protected codeSettings: ICodeSettings;
 
@@ -69,6 +72,7 @@ export default abstract class BaseSnykModule implements IBaseSnykModule {
   protected snykCodeErrorHandler: ISnykCodeErrorHandler;
 
   protected markdownStringAdapter: IMarkdownStringAdapter;
+  readonly workspaceTrust: IWorkspaceTrust;
 
   constructor() {
     this.statusBarItem = new SnykStatusBarItem();
@@ -90,6 +94,7 @@ export default abstract class BaseSnykModule implements IBaseSnykModule {
     this.codeSettings = new CodeSettings(this.snykApiClient, this.contextService, configuration, this.openerService);
     this.advisorApiClient = new AdvisorApiClient(configuration, Logger);
     this.markdownStringAdapter = new MarkdownStringAdapter();
+    this.workspaceTrust = new WorkspaceTrust();
   }
 
   abstract runScan(): Promise<void>;