+++ src/snyk/common/configuration/trustedFolders.ts
@@ -2,12 +2,33 @@ import { IConfiguration } from './configuration';
 
 import path from 'path';
 
-export function getTrustedFolders(config: IConfiguration, workspaceFolders: string[]): string[] {
-  const trustedFolders = config.getTrustedFolders();
-  return workspaceFolders.filter(folder => {
-    return trustedFolders.some(trustedFolder => {
-      const relative = path.relative(trustedFolder, folder);
-      return relative === '' || (relative && !relative.startsWith('..' + path.sep) && !path.isAbsolute(relative));
+export interface IWorkspaceTrust {
+  getTrustedFolders(config: IConfiguration, workspaceFolders: string[]): ReadonlyArray<string>;
+
+  resetTrustedFoldersCache(): void;
+}
+
+export class WorkspaceTrust implements IWorkspaceTrust {
+  private trustedFoldersCache?: ReadonlyArray<string>;
+
+  getTrustedFolders(config: IConfiguration, workspaceFolders: string[]): ReadonlyArray<string> {
+    if (this.trustedFoldersCache !== undefined) {
+      return this.trustedFoldersCache;
+    }
+
+    const trustedFolders = config.getTrustedFolders();
+    const trusted = workspaceFolders.filter(folder => {
+      return trustedFolders.some(trustedFolder => {
+        const relative = path.relative(trustedFolder, folder);
+        return relative === '' || (relative && !relative.startsWith('..' + path.sep) && !path.isAbsolute(relative));
+      });
     });
-  });
+
+    this.trustedFoldersCache = trusted;
+    return trusted;
+  }
+
+  resetTrustedFoldersCache() {
+    this.trustedFoldersCache = undefined;
+  }
 }